fix CPU usage.
a zmq socket is edge-triggered not level-triggered which make it
difficult to use with an eventloop. Previously we were polling on read
and write events which was working but was using 100% of the CPU due to
the fact that the fd is always writable or at least most of the time.
Now we only use the fd to know if the state changed. Due to this change
uzmq.Poll and uzmq.ZMQ behaviours have changed:

- uzmq.Poll is now based on a timer: when time out we are polling the
socket to know if its state changed and the call the callback.
- uzmq.ZMQ is more efficient than uzmq.Poll: we are polling the file
  descriptor to know if the state changed. Also when a message is sent
or received we are preparing a new event to make sure we gathered all
sockets events. This event will happen on next io loop iteration.